-이주연 < 카테고리별 상품 조회 > 1. 데이터베이스 변경 - 상품 db 데이터 삽입
2. serlvet 작성
- index.jsp -> category.jsp 이동 서블릿 (카테고리 목록 가져오기, 상품목록 가져오기)
- 카테고리 목록 가져오기
- css

< 상품 상세조회 >
1. 화면ui 구성
2. productDetail 화면이동
- 상품 객체
- 카테고리별 상품조회 List (관련상품)
- 상품후기

+++ src/com/myHome/model/dto/Product.java
@@ -24,7 +24,7 @@ public class Product implements Serializable {
 	/** 카테고리 번호 */
 	private int categoryId;
 	/** 평점 */
-	private int pScore;
+	private float pScore;
 	/** 판매량 */
 	private int pSales;
 	/** 재고 */
@@ -48,7 +48,7 @@ public class Product implements Serializable {
 	 * @param pCount
 	 */
 	public Product(int pNo, String pName, int pPrice, String pImg, String pDescribe, int deliveryFee,
-			String companyName, int categoryId, int pScore, int pSales, int pCount) {
+			String companyName, int categoryId, float pScore, int pSales, int pCount) {
 		this.pNo = pNo;
 		this.pName = pName;
 		this.pPrice = pPrice;
@@ -127,11 +127,11 @@ public class Product implements Serializable {
 		this.categoryId = categoryId;
 	}
 
-	public int getpScore() {
+	public float getpScore() {
 		return pScore;
 	}
 
-	public void setpScore(int pScore) {
+	public void setpScore(float pScore) {
 		this.pScore = pScore;
 	}
